B2: Fix tests and typings file
